g6

+++ pybricks/pyBricks1/shared.py
@@ -6,9 +6,9 @@ from pybricks.robotics import DriveBase
 class Speed:
     Straight = None
     Slow = float(100)
-    Medium = float(200)
+    Medium = float(250)
     # 600 is max, but sometimes gyro make mistakes so, I guess 500 makes little mistake so 450 is sweet spot
-    Fast = float(300)
+    Fast = float(400)
 
 
 # inspiration https://github.com/FLL-Team-24277/FLL-Fall-2023-Masterpiece/blob/main/samples/base_robot.py
